==> notebooks/01_mcg_preparacion_datos.Rmd <==
---
title: "Preparacion de datos"
author: "Miguel Coto Garcia"
output: html_notebook
---

Cargando paquetes
```{r message=FALSE, warning=FALSE, include=FALSE}

library(here)
library(tidyverse)

```


# Rutas
```{r}

raw_data <- here("data", "raw")
interim_data <- here("data", "interim")
processed_data <- here("data", "processed")

```

# Transformando datos

## Leyendo bases
```{r}

base <- readxl::read_excel(paste0(raw_data,"/BASE TARJETAS DE CREDITO.xls"), sheet = "Base") %>% 
  janitor::clean_names()
nuevos <- readxl::read_excel(paste0(raw_data,"/BASE TARJETAS DE CREDITO.xls"), sheet = "Casos nuevos")%>% 
  janitor::clean_names()
profesiones <- readxl::read_excel(paste0(raw_data,"/BASE TARJETAS DE CREDITO.xls"), sheet = "profesiones")%>% 
  janitor::clean_names()
estado_civil <- readxl::read_excel(paste0(raw_data,"/BASE TARJETAS DE CREDITO.xls"), sheet = "estado_civil")%>% 
  janitor::clean_names()
ingresos <- readxl::read_excel(paste0(raw_data,"/BASE TARJETAS DE CREDITO.xls"), sheet = "tipo_ingresos")%>% 
  janitor::clean_names()

```


## Asignando categorias de profesion y de ingresos
```{r}

base_final <- base %>%
  left_join(ingresos, by = "codg_sector_trab") %>%
  left_join(profesiones, by = "codigo_profesion") %>%
  select(
    -c(
      tipo_ingresos,
      profesion,
      profesion2,
      profesion3,
      plazo_meses,
      destino,
      codigo_profesion,
      codg_sector_trab
    )
  ) %>%
  mutate(buen_pagador = if_else(buen_pagador == "bueno", 0, 1)) %>%
  rename(tipo_ingresos = tipo_ingresos_2,
         profesion = profesion4,
         objetivo = buen_pagador
         ) %>%
  mutate(across(where(is.character), as.factor))

head(base_final)
```
```{r}
nuevos_final <- nuevos %>%
  left_join(ingresos, by = "codg_sector_trab") %>%
  left_join(profesiones, by = "codigo_profesion") %>%
  select(
    -c(
      tipo_ingresos,
      profesion,
      profesion2,
      profesion3,
      plazo_meses,
      destino,
      codigo_profesion,
      codg_sector_trab
    )
  ) %>%
  rename(tipo_ingresos = tipo_ingresos_2,
         profesion = profesion4,
         objetivo = buen_pagador) %>%
  mutate(across(where(is.character), as.factor))

head(nuevos_final)

```



```{r}
saveRDS(base_final, file=paste0(processed_data,"/base_final.Rds"))
saveRDS(nuevos_final, file=paste0(processed_data,"/nuevos_final.Rds"))

```
