==> notebooks/03_mcg_benchmark.Rmd <==
---
title: "Benchmark de modelo"
author: "Miguel Coto Garcia"
output: html_notebook
---

Cargando paquetes
```{r message=FALSE, warning=FALSE, include=FALSE}

library(here)
library(tidyverse)
library(mlr3)
library("mlr3viz")
library("mlr3learners")
library(mlr3pipelines)
#library(mlr3extralearners)
library("mlr3tuning")
```

# Rutas
```{r}

raw_data <- here("data", "raw")
interim_data <- here("data", "interim")
processed_data <- here("data", "processed")

```

## Leyendo bases
```{r}
base <- readRDS(file=paste0(processed_data,"/base_final.Rds"))

base <- base %>%
  mutate(objetivo = factor(if_else(objetivo==0, "bueno", "malo")))

head(base)
```

# Tarea de aprendizaje
## Definicion



```{r}
task_tarj = TaskClassif$new(id = "tarjetas", backend = base, target = "objetivo")
print(task_tarj)
```

## Separacion de training y testing
```{r}
set.seed(34678)
train_set = sample(task_tarj$nrow, 0.8 * task_tarj$nrow)
test_set = setdiff(seq_len(task_tarj$nrow), train_set)
```




```{r message=FALSE, warning=FALSE}
#autoplot(task_tarj$select(c("saldo_tarjeta", "coeficiente_solvencia", "edad", "limite_tarjeta_credito")), type = "pairs")
#task_tarj = TaskClassif$new(id = "tarjetas", backend = base, target = "objetivo")
```

# Modelos

## Lista de modelos

```{r}

lrn_rpart = lrn("classif.rpart", predict_type = "prob")
lrn_glmnet = lrn("classif.glmnet", predict_type = "prob")
lrn_knn = lrn("classif.kknn", predict_type = "prob")
lrn_lda = lrn("classif.lda", predict_type = "prob")
lrn_nnet = lrn("classif.nnet", predict_type = "prob")
lrn_rf = lrn("classif.ranger", predict_type = "prob")
lrn_svm = lrn("classif.svm", predict_type = "prob")
lrn_xgboost = lrn("classif.xgboost", predict_type = "prob")

```

# Benchmark

## Prepracion de datos para el benchmark
### Punto de corte: 0.5
```{r}
#th = prop.table(table(base$objetivo))[1]
th = .5
impute_fcts <- po("imputemode", affect_columns = selector_type("factor"))
impute_nums <- po("imputehist", affect_columns = selector_type("numeric"))
encode <- po("encode", affect_columns = selector_type("factor"))
threshold <- po("threshold", param_vals = list(thresholds = th))

pre_procesamiento <- impute_fcts %>>% 
  impute_nums %>>% 
  encode

lrn_rpart$predict_type = "prob"
lrn_glmnet$predict_type = "prob"
lrn_knn$predict_type = "prob"
lrn_lda$predict_type = "prob"
lrn_nnet$predict_type = "prob"
lrn_rf$predict_type = "prob"
lrn_svm$predict_type = "prob"
lrn_xgboost$predict_type = "prob"

lrn_rpart <- GraphLearner$new(pre_procesamiento %>>% po(lrn_rpart) %>>% threshold) 
lrn_glmnet <- GraphLearner$new(pre_procesamiento %>>% po(lrn_glmnet) %>>% threshold) 
lrn_knn <- GraphLearner$new(pre_procesamiento %>>% po(lrn_knn) %>>% threshold) 
lrn_lda <- GraphLearner$new(pre_procesamiento %>>% po(lrn_lda) %>>% threshold) 
lrn_nnet <- GraphLearner$new(pre_procesamiento %>>% po(lrn_nnet) %>>% threshold) 
lrn_rf <- GraphLearner$new(pre_procesamiento %>>% po(lrn_rf) %>>% threshold) 
lrn_svm <- GraphLearner$new(pre_procesamiento %>>% po(lrn_svm) %>>% threshold) 
lrn_xgboost <- GraphLearner$new(pre_procesamiento %>>% po(lrn_xgboost) %>>% threshold) 

lrn_rpart$predict_sets = c("train", "test")
lrn_glmnet$predict_sets = c("train", "test")
lrn_knn$predict_sets = c("train", "test")
lrn_lda$predict_sets = c("train", "test")
#lrn_log_reg$predict_sets = c("train", "test")
lrn_nnet$predict_sets = c("train", "test")
lrn_rf$predict_sets = c("train", "test")
lrn_svm$predict_sets = c("train", "test")
lrn_xgboost$predict_sets = c("train", "test")

lrn_rpart$id = "Árbol"
lrn_glmnet$id = "Reg-reg"
lrn_knn$id = "K-vecinos"
lrn_lda$id = "LDA"
#lrn_log_reg$id = "Reg-log"
lrn_nnet$id = "Red-Neur"
lrn_rf$id = "RandomForest"
lrn_svm$id = "SVM"
lrn_xgboost$id = "XGBoost"

learners = list(
  lrn_rpart,
  lrn_glmnet,
  lrn_knn,
  lrn_lda,
  #lrn_log_reg,
  lrn_nnet,
  lrn_rf,
  lrn_svm,
  lrn_xgboost
)


```

## Defincion del benchmark
```{r}
resamplings = rsmp("cv", folds = 10)

design = benchmark_grid(task_tarj, learners, resamplings)

```


## Ejecucion
```{r message=FALSE, warning=FALSE}
future::plan("multiprocess")

bmr = benchmark(design)
```

## Medidas

```{r}
costs = matrix(c(0, 2, 3, -1), nrow = 2)
dimnames(costs) = list(response = c("bueno", "malo"), truth = c("bueno", "malo"))
cost_measure = msr("classif.costs", costs = costs)

```


```{r}
measures = list(
  msr("classif.acc", id = "acc_train", predict_sets = "train"),
  msr("classif.acc", id = "acc_test"),
  msr("classif.ppv", id = "ppv_train", predict_sets = "train"),
  msr("classif.ppv", id = "ppv_test"),
  msr("classif.tpr", id = "tpr_train", predict_sets = "train"),
  msr("classif.tpr", id = "tpr_test"),
  msr("classif.costs", costs = costs, id = "cost_train", predict_sets = "train"),
  msr("classif.costs", costs = costs, id = "cost_test")
)

bmr$aggregate(measures)

```


```{r}

tab = bmr$aggregate(measures)

ranks = tab[, .(learner_id, rank_train = rank(-acc_train), rank_test = rank(-acc_test)), by = task_id]
print(ranks)
```

## Graficos comparativos


```{r}
autoplot(bmr$clone(deep = TRUE), type = "roc")

```


### Area bajo la curva de ROC
```{r}
autoplot(bmr, measure = msr("classif.auc")) + 
           theme(axis.text.x = element_text(angle = 45, hjust = 1))


```


### Precision
```{r}
autoplot(bmr, measure = msr("classif.acc")) + 
           theme(axis.text.x = element_text(angle = 45, hjust = 1))


```

### Sensitividad (tpr)
```{r}
autoplot(bmr, measure = msr("classif.tpr")) + 
           theme(axis.text.x = element_text(angle = 45, hjust = 1))


```

### Precicsion (ppv)
```{r}
autoplot(bmr, measure = msr("classif.ppv")) + 
           theme(axis.text.x = element_text(angle = 45, hjust = 1))


```
### Costo
```{r}
autoplot(bmr, measure = cost_measure) + 
           theme(axis.text.x = element_text(angle = 45, hjust = 1))


```

### Punto de corte: 0.92
```{r}
th2 = prop.table(table(base$objetivo))[1]

threshold2 <- po("threshold", param_vals = list(thresholds = th2))

lrn_rpart = lrn("classif.rpart", predict_type = "prob")
lrn_glmnet = lrn("classif.glmnet", predict_type = "prob")
lrn_knn = lrn("classif.kknn", predict_type = "prob")
lrn_lda = lrn("classif.lda", predict_type = "prob")
lrn_nnet = lrn("classif.nnet", predict_type = "prob")
lrn_rf = lrn("classif.ranger", predict_type = "prob")
lrn_svm = lrn("classif.svm", predict_type = "prob")
lrn_xgboost = lrn("classif.xgboost", predict_type = "prob")

lrn_rpart <- GraphLearner$new(pre_procesamiento %>>% po(lrn_rpart) %>>% threshold2) 
lrn_glmnet <- GraphLearner$new(pre_procesamiento %>>% po(lrn_glmnet) %>>% threshold2) 
lrn_knn <- GraphLearner$new(pre_procesamiento %>>% po(lrn_knn) %>>% threshold2) 
lrn_lda <- GraphLearner$new(pre_procesamiento %>>% po(lrn_lda) %>>% threshold2) 
lrn_nnet <- GraphLearner$new(pre_procesamiento %>>% po(lrn_nnet) %>>% threshold2) 
lrn_rf <- GraphLearner$new(pre_procesamiento %>>% po(lrn_rf) %>>% threshold2) 
lrn_svm <- GraphLearner$new(pre_procesamiento %>>% po(lrn_svm) %>>% threshold2) 
lrn_xgboost <- GraphLearner$new(pre_procesamiento %>>% po(lrn_xgboost) %>>% threshold2) 

lrn_rpart$predict_sets = c("train", "test")
lrn_glmnet$predict_sets = c("train", "test")
lrn_knn$predict_sets = c("train", "test")
lrn_lda$predict_sets = c("train", "test")
#lrn_log_reg$predict_sets = c("train", "test")
lrn_nnet$predict_sets = c("train", "test")
lrn_rf$predict_sets = c("train", "test")
lrn_svm$predict_sets = c("train", "test")
lrn_xgboost$predict_sets = c("train", "test")

lrn_rpart$id = "Árbol"
lrn_glmnet$id = "Reg-reg"
lrn_knn$id = "K-vecinos"
lrn_lda$id = "LDA"
#lrn_log_reg$id = "Reg-log"
lrn_nnet$id = "Red-Neur"
lrn_rf$id = "RandomForest"
lrn_svm$id = "SVM"
lrn_xgboost$id = "XGBoost"

learners2 = list(
  lrn_rpart,
  lrn_glmnet,
  lrn_knn,
  lrn_lda,
  #lrn_log_reg,
  lrn_nnet,
  lrn_rf,
  lrn_svm,
  lrn_xgboost
)


```

## Defincion del benchmark
```{r}
design2 = benchmark_grid(task_tarj, learners2, resamplings)

```


## Ejecucion
```{r message=FALSE, warning=FALSE}
future::plan("multiprocess")

bmr2 = benchmark(design2)
```

## Medidas


```{r}

bmr2$aggregate(measures)

```


```{r}

tab2 = bmr2$aggregate(measures)

ranks2 = tab2[, .(learner_id, rank_train = rank(-acc_train), rank_test = rank(-acc_test)), by = task_id]
print(ranks2)
```

## Graficos comparativos


```{r}
autoplot(bmr2$clone(deep = TRUE), type = "roc")

```


### Area bajo la curva de ROC
```{r}
autoplot(bmr2, measure = msr("classif.auc")) + 
           theme(axis.text.x = element_text(angle = 45, hjust = 1))


```


### Precision
```{r}
autoplot(bmr2, measure = msr("classif.acc")) + 
           theme(axis.text.x = element_text(angle = 45, hjust = 1))


```

### Sensitividad (tpr)
```{r}
autoplot(bmr2, measure = msr("classif.tpr")) + 
           theme(axis.text.x = element_text(angle = 45, hjust = 1))


```

### Precicsion (ppv)
```{r}
autoplot(bmr2, measure = msr("classif.ppv")) + 
           theme(axis.text.x = element_text(angle = 45, hjust = 1))


```

### Costo
```{r}
autoplot(bmr2, measure = cost_measure) + 
           theme(axis.text.x = element_text(angle = 45, hjust = 1))


```

```{r}

# lrn_rf_2 = lrn_rf
# 
# lrn_rf_2$train(task_tarj, row_ids = train_set)
# pred_fr = lrn_rf_2$predict(task_tarj, row_ids = test_set)
# 
# pred_fr$confusion

```

```{r}

# lrn_xgboost_2 = lrn_xgboost
# 
# lrn_xgboost_2$train(task_tarj, row_ids = train_set)
# pred_xg = lrn_xgboost_2$predict(task_tarj, row_ids = test_set, )
# 
# pred_xg$confusion


```
